Properly returns errors on trpc.register
